refactor(list): add index in web and native folders and replace template string way, at styleds
