+++ package.json
@@ -1,6 +1,7 @@
 {
   "scripts": {
     "commit": "git-cz",
+    "clean": "lerna clean -y && lerna run prebuild --scope=@gympass/yoga-doc",
     "test": "NODE_ENV=test jest",
     "pretest": "lerna run prebuild",
     "test:native": "NODE_ENV=test jest --config=jest.config.native.js",
