+++ packages/yoga/src/List/web/List.jsx
@@ -2,38 +2,37 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 
-const StyledList = styled.ul`
-  ${({
+const StyledList = styled.ul(
+  ({
     divided,
     horizontal,
     theme: {
-      components: {
-        list: {
-          border: { width: borderWidth, color: borderColor },
-        },
-      },
+      components: { list },
     },
   }) => `
-    display: flex;
-    width: 100%;
-    flex-direction: ${horizontal ? 'row' : 'column'};
+  display: flex;
+  width: 100%;
+  flex-direction: ${horizontal ? 'row' : 'column'};
+  list-style: none;
+  padding: 0;
+  margin: 0;
 
-    > * {
-        ${
-          divided
-            ? `border-${
-                horizontal ? 'right' : 'bottom'
-              }: ${borderWidth}px solid ${borderColor}`
-            : ''
-        };
+  > * {
+      ${
+        divided
+          ? `border-${horizontal ? 'right' : 'bottom'}: ${
+              list.border.width
+            }px solid ${list.border.color}`
+          : ''
+      };
 
-        &:last-child {
-          border: none;
-        }
+      &:last-child {
+        border: none;
       }
     }
-  `}
-`;
+  }
+`,
+);
 
 /** Lists are a continuous group of text or images. They are composed of items containing primary and supplemental actions, which are represented by icons and text. */
 const List = ({ horizontal, divided, ...rest }) => (
