+++ packages/yoga/src/List/web/index.js
@@ -1,0 +1,4 @@
+import ListItem from './ListItem';
+import List from './List';
+
+export { List, ListItem };
